REMOVED OLD LICENSE FROM collectr.cc
profile: UML, SBOM

- removed old license from collectr.cc in 2007.

LICENSE NOTE:
collectr.cc was a project back in the day by the committer. Some part of the code / generator templates was carried over to this project. The author of this commit is the copyright and license giver. Therefore, removal of the license comes from the source of copyright. Nothing to freak out about. ;-)